@@ -1,6 +1,6 @@
 module github.com/nsqio/nsq
 
-go 1.16
+go 1.17
 
 require (
 	github.com/BurntSushi/toml v0.4.1
