@@ -39,30 +39,38 @@ func (t *tlsRequiredOption) IsBoolFlag() bool { return true }
 
 type tlsMinVersionOption uint16
 
+var tlsVersionTable = []struct {
+	val uint16
+	str string
+}{
+	{tls.VersionTLS10, "tls1.0"},
+	{tls.VersionTLS11, "tls1.1"},
+	{tls.VersionTLS12, "tls1.2"},
+	{tls.VersionTLS13, "tls1.3"},
+}
+
 func (t *tlsMinVersionOption) Set(s string) error {
 	s = strings.ToLower(s)
-	switch s {
-	case "":
+	if s == "" {
 		return nil
-	case "ssl3.0":
-		*t = tls.VersionSSL30
-	case "tls1.0":
-		*t = tls.VersionTLS10
-	case "tls1.1":
-		*t = tls.VersionTLS11
-	case "tls1.2":
-		*t = tls.VersionTLS12
-	case "tls1.3":
-		*t = tls.VersionTLS13
-	default:
-		return fmt.Errorf("unknown tlsVersionOption %q", s)
 	}
-	return nil
+	for _, v := range tlsVersionTable {
+		if s == v.str {
+			*t = tlsMinVersionOption(v.val)
+			return nil
+		}
+	}
+	return fmt.Errorf("unknown tlsVersionOption %q", s)
 }
 
 func (t *tlsMinVersionOption) Get() interface{} { return uint16(*t) }
 
 func (t *tlsMinVersionOption) String() string {
+	for _, v := range tlsVersionTable {
+		if uint16(*t) == v.val {
+			return v.str
+		}
+	}
 	return strconv.FormatInt(int64(*t), 10)
 }
 
@@ -121,8 +129,9 @@ func nsqdFlagSet(opts *nsqd.Options) *flag.FlagSet {
 	flagSet.Bool("worker-id", false, "[deprecated] use --node-id")
 
 	flagSet.String("https-address", opts.HTTPSAddress, "<addr>:<port> to listen on for HTTPS clients")
-	flagSet.String("http-address", opts.HTTPAddress, "<addr>:<port> to listen on for HTTP clients")
-	flagSet.String("tcp-address", opts.TCPAddress, "<addr>:<port> to listen on for TCP clients")
+	flagSet.String("http-address", opts.HTTPAddress, "address to listen on for HTTP clients (<addr>:<port> for TCP/IP or <path> for unix socket)")
+	flagSet.String("tcp-address", opts.TCPAddress, "address to listen on for TCP clients (<addr>:<port> for TCP/IP or <path> for unix socket)")
+
 	authHTTPAddresses := app.StringArray{}
 	flagSet.Var(&authHTTPAddresses, "auth-http-address", "<addr>:<port> or a full url to query auth server (may be given multiple times)")
 	flagSet.String("broadcast-address", opts.BroadcastAddress, "address that will be registered with lookupd (defaults to the OS hostname)")
