@@ -65,15 +65,15 @@ func (m *Message) WriteTo(w io.Writer) (int64, error) {
 }
 
 // decodeMessage deserializes data (as []byte) and creates a new Message
-// message format:
-// [x][x][x][x][x][x][x][x][x][x][x][x][x][x][x][x][x][x][x][x][x][x][x][x][x][x][x][x][x][x]...
-// |       (int64)        ||    ||      (hex string encoded in ASCII)           || (binary)
-// |       8-byte         ||    ||                 16-byte                      || N-byte
-// ------------------------------------------------------------------------------------------...
-//   nanosecond timestamp    ^^                   message ID                       message body
-//                        (uint16)
-//                         2-byte
-//                        attempts
+//
+//	[x][x][x][x][x][x][x][x][x][x][x][x][x][x][x][x][x][x][x][x][x][x][x][x][x][x][x][x][x][x]...
+//	|       (int64)        ||    ||      (hex string encoded in ASCII)           || (binary)
+//	|       8-byte         ||    ||                 16-byte                      || N-byte
+//	------------------------------------------------------------------------------------------...
+//	  nanosecond timestamp    ^^                   message ID                       message body
+//	                       (uint16)
+//	                        2-byte
+//	                       attempts
 func decodeMessage(b []byte) (*Message, error) {
 	var msg Message
 
